TWINS-234: moved FactoryConditionSetRestDTOMapper in correctly folder and change version FactoryPipelineDTOv2 in RelatedObjects
