+++ core/src/main/java/org/twins/core/dto/rest/related/RelatedObjectsDTOv1.java
@@ -8,6 +8,7 @@ import org.twins.core.dto.rest.datalist.DataListOptionDTOv1;
 import org.twins.core.dto.rest.businessaccount.BusinessAccountDTOv1;
 import org.twins.core.dto.rest.factory.FactoryDTOv1;
 import org.twins.core.dto.rest.factory.FactoryPipelineDTOv1;
+import org.twins.core.dto.rest.factory.FactoryPipelineDTOv2;
 import org.twins.core.dto.rest.featurer.FeaturerDTOv1;
 import org.twins.core.dto.rest.permission.PermissionDTOv1;
 import org.twins.core.dto.rest.permission.PermissionGroupDTOv1;
@@ -74,7 +75,7 @@ public class RelatedObjectsDTOv1 {
     public Map<UUID, FactoryDTOv1> factoryMap;
 
     @Schema(description = "related factory pipeline map", example = "{factory pipeline map}")
-    public Map<UUID, FactoryPipelineDTOv1> factoryPipelineMap;
+    public Map<UUID, FactoryPipelineDTOv2> factoryPipelineMap;
 
     @Schema(description = "related featurer map", example = "{featurer map}")
     public Map<Integer, FeaturerDTOv1> featurerMap;
