+++ core/src/main/java/org/twins/core/mappers/rest/factory/FactoryBranchRestDTOMapperV2.java
@@ -9,7 +9,6 @@ import org.twins.core.mappers.rest.RestSimpleDTOMapper;
 import org.twins.core.mappers.rest.mappercontext.MapperContext;
 import org.twins.core.mappers.rest.mappercontext.modes.FactoryConditionSetMode;
 import org.twins.core.mappers.rest.mappercontext.modes.FactoryMode;
-import org.twins.core.mappers.rest.permission.FactoryConditionSetRestDTOMapper;
 
 @Component
 @RequiredArgsConstructor
