+++ core/src/main/java/org/twins/core/mappers/rest/factory/FactoryPipelineRestDTOMapperV2.java
@@ -8,7 +8,6 @@ import org.twins.core.dto.rest.factory.FactoryPipelineDTOv2;
 import org.twins.core.mappers.rest.RestSimpleDTOMapper;
 import org.twins.core.mappers.rest.mappercontext.MapperContext;
 import org.twins.core.mappers.rest.mappercontext.modes.*;
-import org.twins.core.mappers.rest.permission.FactoryConditionSetRestDTOMapper;
 import org.twins.core.mappers.rest.twinclass.TwinClassBaseRestDTOMapper;
 import org.twins.core.mappers.rest.twinstatus.TwinStatusRestDTOMapper;
